refactor: slot manager tests
